+++ config/application.rb
@@ -17,6 +17,8 @@ require_relative "../lib/dfe_sign_in"
 require_relative "../lib/dqt"
 require_relative "../lib/ordnance_survey"
 require_relative "../lib/one_time_password"
+require_relative "../lib/send_analytics_csv"
+require_relative "../lib/upload"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.